requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=2.5.0
opencv-python
h5py
imgaug
IPython[all]
